refactor: 이용시간 기록시 백그라운드 진입시간과 현재시간 차이를 반영하도록 수정

+++ HongikYeolgong2-iOS/Scene/ReadingRoom/Calendar/CalendarCell.swift
@@ -18,7 +18,7 @@ enum CellStyle: CaseIterable {
 struct CalendarCell: View {
     let dayInfo: Day
     
-    private var totalTime: Int {
+    private var totalTime: Double {
         if let stduyRecords = dayInfo.usageRecord {
             return stduyRecords.map { $0.duration }.reduce(0, +)
         } else {
